+++ clang/lib/Headers/avx512vldqintrin.h
@@ -17,12 +17,10 @@
 /* Define the default attributes for the functions in this file. */
 #define __DEFAULT_FN_ATTRS128                                                  \
   __attribute__((__always_inline__, __nodebug__,                               \
-                 __target__("avx512vl,avx512dq,no-evex512"),                   \
-                 __min_vector_width__(128)))
+                 __target__("avx512vl,avx512dq"), __min_vector_width__(128)))
 #define __DEFAULT_FN_ATTRS256                                                  \
   __attribute__((__always_inline__, __nodebug__,                               \
-                 __target__("avx512vl,avx512dq,no-evex512"),                   \
-                 __min_vector_width__(256)))
+                 __target__("avx512vl,avx512dq"), __min_vector_width__(256)))
 
 #if defined(__cplusplus) && (__cplusplus >= 201103L)
 #define __DEFAULT_FN_ATTRS256_CONSTEXPR __DEFAULT_FN_ATTRS256 constexpr