+++ clang/test/CodeGen/target-builtin-noerror.c
@@ -141,10 +141,8 @@ void verifyfeaturestrings(void) {
   (void)__builtin_cpu_supports("sm4");
   (void)__builtin_cpu_supports("apxf");
   (void)__builtin_cpu_supports("usermsr");
-  (void)__builtin_cpu_supports("avx10.1-256");
-  (void)__builtin_cpu_supports("avx10.1-512");
-  (void)__builtin_cpu_supports("avx10.2-256");
-  (void)__builtin_cpu_supports("avx10.2-512");
+  (void)__builtin_cpu_supports("avx10.1");
+  (void)__builtin_cpu_supports("avx10.2");
   (void)__builtin_cpu_supports("movrs");
 }